+++ sdk/resourcemanager/attestation/armattestation/providers_client.go
@@ -27,7 +27,7 @@ type ProvidersClient struct {
 // NewProvidersClient creates a new instance of ProvidersClient with the specified values.
 //   - subscriptionID - The ID of the target subscription.
 //   - credential - used to authorize requests. Usually a credential from azidentity.
-//   - options - pass nil to accept the default values.
+//   - options - Contains optional client configuration. Pass nil to accept the default values.
 func NewProvidersClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*ProvidersClient, error) {
 	cl, err := arm.NewClient(moduleName, moduleVersion, credential, options)
 	if err != nil {
@@ -40,10 +40,10 @@ func NewProvidersClient(subscriptionID string, credential azcore.TokenCredential
 	return client, nil
 }
 
-// Create - Creates a new Attestation Provider.
+// Create - Creates or updates an Attestation Provider.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-10-01
+// Generated from API version 2021-06-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - providerName - Name of the attestation provider.
 //   - creationParams - Client supplied parameters.
@@ -71,7 +71,7 @@ func (client *ProvidersClient) Create(ctx context.Context, resourceGroupName str
 }
 
 // createCreateRequest creates the Create request.
-func (client *ProvidersClient) createCreateRequest(ctx context.Context, resourceGroupName string, providerName string, creationParams ServiceCreationParams, options *ProvidersClientCreateOptions) (*policy.Request, error) {
+func (client *ProvidersClient) createCreateRequest(ctx context.Context, resourceGroupName string, providerName string, creationParams ServiceCreationParams, _ *ProvidersClientCreateOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Attestation/attestationProviders/{providerName}"
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
@@ -90,7 +90,7 @@ func (client *ProvidersClient) createCreateRequest(ctx context.Context, resource
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-10-01")
+	reqQP.Set("api-version", "2021-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, creationParams); err != nil {
@@ -111,7 +111,7 @@ func (client *ProvidersClient) createHandleResponse(resp *http.Response) (Provid
 // Delete - Delete Attestation Service.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-10-01
+// Generated from API version 2021-06-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - providerName - Name of the attestation service
 //   - options - ProvidersClientDeleteOptions contains the optional parameters for the ProvidersClient.Delete method.
@@ -137,7 +137,7 @@ func (client *ProvidersClient) Delete(ctx context.Context, resourceGroupName str
 }
 
 // deleteCreateRequest creates the Delete request.
-func (client *ProvidersClient) deleteCreateRequest(ctx context.Context, resourceGroupName string, providerName string, options *ProvidersClientDeleteOptions) (*policy.Request, error) {
+func (client *ProvidersClient) deleteCreateRequest(ctx context.Context, resourceGroupName string, providerName string, _ *ProvidersClientDeleteOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Attestation/attestationProviders/{providerName}"
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
@@ -156,7 +156,7 @@ func (client *ProvidersClient) deleteCreateRequest(ctx context.Context, resource
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-10-01")
+	reqQP.Set("api-version", "2021-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -165,7 +165,7 @@ func (client *ProvidersClient) deleteCreateRequest(ctx context.Context, resource
 // Get - Get the status of Attestation Provider.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-10-01
+// Generated from API version 2021-06-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - providerName - Name of the attestation provider.
 //   - options - ProvidersClientGetOptions contains the optional parameters for the ProvidersClient.Get method.
@@ -192,7 +192,7 @@ func (client *ProvidersClient) Get(ctx context.Context, resourceGroupName string
 }
 
 // getCreateRequest creates the Get request.
-func (client *ProvidersClient) getCreateRequest(ctx context.Context, resourceGroupName string, providerName string, options *ProvidersClientGetOptions) (*policy.Request, error) {
+func (client *ProvidersClient) getCreateRequest(ctx context.Context, resourceGroupName string, providerName string, _ *ProvidersClientGetOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Attestation/attestationProviders/{providerName}"
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
@@ -211,7 +211,7 @@ func (client *ProvidersClient) getCreateRequest(ctx context.Context, resourceGro
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-10-01")
+	reqQP.Set("api-version", "2021-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -229,7 +229,7 @@ func (client *ProvidersClient) getHandleResponse(resp *http.Response) (Providers
 // GetDefaultByLocation - Get the default provider by location.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-10-01
+// Generated from API version 2021-06-01
 //   - location - The location of the default provider.
 //   - options - ProvidersClientGetDefaultByLocationOptions contains the optional parameters for the ProvidersClient.GetDefaultByLocation
 //     method.
@@ -256,7 +256,7 @@ func (client *ProvidersClient) GetDefaultByLocation(ctx context.Context, locatio
 }
 
 // getDefaultByLocationCreateRequest creates the GetDefaultByLocation request.
-func (client *ProvidersClient) getDefaultByLocationCreateRequest(ctx context.Context, location string, options *ProvidersClientGetDefaultByLocationOptions) (*policy.Request, error) {
+func (client *ProvidersClient) getDefaultByLocationCreateRequest(ctx context.Context, location string, _ *ProvidersClientGetDefaultByLocationOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Attestation/locations/{location}/defaultProvider"
 	if location == "" {
 		return nil, errors.New("parameter location cannot be empty")
@@ -271,7 +271,7 @@ func (client *ProvidersClient) getDefaultByLocationCreateRequest(ctx context.Con
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-10-01")
+	reqQP.Set("api-version", "2021-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -289,7 +289,7 @@ func (client *ProvidersClient) getDefaultByLocationHandleResponse(resp *http.Res
 // List - Returns a list of attestation providers in a subscription.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-10-01
+// Generated from API version 2021-06-01
 //   - options - ProvidersClientListOptions contains the optional parameters for the ProvidersClient.List method.
 func (client *ProvidersClient) List(ctx context.Context, options *ProvidersClientListOptions) (ProvidersClientListResponse, error) {
 	var err error
@@ -314,7 +314,7 @@ func (client *ProvidersClient) List(ctx context.Context, options *ProvidersClien
 }
 
 // listCreateRequest creates the List request.
-func (client *ProvidersClient) listCreateRequest(ctx context.Context, options *ProvidersClientListOptions) (*policy.Request, error) {
+func (client *ProvidersClient) listCreateRequest(ctx context.Context, _ *ProvidersClientListOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Attestation/attestationProviders"
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
@@ -325,7 +325,7 @@ func (client *ProvidersClient) listCreateRequest(ctx context.Context, options *P
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-10-01")
+	reqQP.Set("api-version", "2021-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -343,7 +343,7 @@ func (client *ProvidersClient) listHandleResponse(resp *http.Response) (Provider
 // ListByResourceGroup - Returns attestation providers list in a resource group.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-10-01
+// Generated from API version 2021-06-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - options - ProvidersClientListByResourceGroupOptions contains the optional parameters for the ProvidersClient.ListByResourceGroup
 //     method.
@@ -370,7 +370,7 @@ func (client *ProvidersClient) ListByResourceGroup(ctx context.Context, resource
 }
 
 // listByResourceGroupCreateRequest creates the ListByResourceGroup request.
-func (client *ProvidersClient) listByResourceGroupCreateRequest(ctx context.Context, resourceGroupName string, options *ProvidersClientListByResourceGroupOptions) (*policy.Request, error) {
+func (client *ProvidersClient) listByResourceGroupCreateRequest(ctx context.Context, resourceGroupName string, _ *ProvidersClientListByResourceGroupOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Attestation/attestationProviders"
 	if resourceGroupName == "" {
 		return nil, errors.New("parameter resourceGroupName cannot be empty")
@@ -385,7 +385,7 @@ func (client *ProvidersClient) listByResourceGroupCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-10-01")
+	reqQP.Set("api-version", "2021-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -403,7 +403,7 @@ func (client *ProvidersClient) listByResourceGroupHandleResponse(resp *http.Resp
 // ListDefault - Get the default provider
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-10-01
+// Generated from API version 2021-06-01
 //   - options - ProvidersClientListDefaultOptions contains the optional parameters for the ProvidersClient.ListDefault method.
 func (client *ProvidersClient) ListDefault(ctx context.Context, options *ProvidersClientListDefaultOptions) (ProvidersClientListDefaultResponse, error) {
 	var err error
@@ -428,7 +428,7 @@ func (client *ProvidersClient) ListDefault(ctx context.Context, options *Provide
 }
 
 // listDefaultCreateRequest creates the ListDefault request.
-func (client *ProvidersClient) listDefaultCreateRequest(ctx context.Context, options *ProvidersClientListDefaultOptions) (*policy.Request, error) {
+func (client *ProvidersClient) listDefaultCreateRequest(ctx context.Context, _ *ProvidersClientListDefaultOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/providers/Microsoft.Attestation/defaultProviders"
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
@@ -439,7 +439,7 @@ func (client *ProvidersClient) listDefaultCreateRequest(ctx context.Context, opt
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-10-01")
+	reqQP.Set("api-version", "2021-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -457,7 +457,7 @@ func (client *ProvidersClient) listDefaultHandleResponse(resp *http.Response) (P
 // Update - Updates the Attestation Provider.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-10-01
+// Generated from API version 2021-06-01
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - providerName - Name of the attestation provider.
 //   - updateParams - Client supplied parameters.
@@ -485,7 +485,7 @@ func (client *ProvidersClient) Update(ctx context.Context, resourceGroupName str
 }
 
 // updateCreateRequest creates the Update request.
-func (client *ProvidersClient) updateCreateRequest(ctx context.Context, resourceGroupName string, providerName string, updateParams ServicePatchParams, options *ProvidersClientUpdateOptions) (*policy.Request, error) {
+func (client *ProvidersClient) updateCreateRequest(ctx context.Context, resourceGroupName string, providerName string, updateParams ServicePatchParams, _ *ProvidersClientUpdateOptions) (*policy.Request, error) {
 	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Attestation/attestationProviders/{providerName}"
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
@@ -504,7 +504,7 @@ func (client *ProvidersClient) updateCreateRequest(ctx context.Context, resource
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-10-01")
+	reqQP.Set("api-version", "2021-06-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, updateParams); err != nil {